+++ drash/vue/pages/tutorials/misc/adding_global_members.vue
@@ -10,73 +10,59 @@ export default {
   data() {
     return {
       example_code: this.$example_code['drash/example_code/tutorials/misc/adding_global_members'],
-      toc: {
-        items: [
-          "Before You Get Started",
-          "Folder Structure End State",
-          "Steps",
-          "Verification",
-        ]
-      }
+      title: resource.meta.title,
+      toc: [
+        "Before You Get Started",
+        "Folder Structure End State",
+        "Steps",
+        "Verification",
+      ]
     };
   },
 }
 </script>
 
 <template lang="pug">
-page-tutorial(
+page(
+  :base_url="this.$base_url"
+  :title="title"
   :toc="toc"
 )
-  div.row
-    div.col
-      hr
-      h2-hash Before You Get Started
-      p A global member is any "thing" you define the global member to be. It can be a service, an app, an array, a number, etc.
-      p You can add global members to the <code>Drash.Members</code> namespace via <code>Drash.addMember()</code>.
-      p Adding global members is helpful when you want to access a "thing" without having to remember where it's physically located. This prevents you from having to import like <code>import MyThing from "../path/../to/../some/../dir/my_thing.ts"</code>. Instead, you can add your "thing" as a global member (e.g., <code>Drash.addMember("MyThing", myThing);</code>) and access it like <code>Drash.Members.MyThing</code>.
-      p All global members are added to the <code>Drash.Members</code> namespace.
-      p-view-source-code(:source_code_uri="$route.meta.source_code_uri")
-  div.row
-    div.col
-      hr
-      div-folder-structure-end-state(:code_block_data="example_code.folder_structure")
-  div.row
-    div.col
-      hr
-      h2-hash Steps
-      ol
-        li
-          p Create your "thing" to be added as a global member.
-          p
-            code-block-slotted(language="typescript")
-              template(v-slot:title) {{ example_code.my_thing.filepath }}
-              template(v-slot:code) {{ example_code.my_thing.contents }}
-        li
-          p Create a file that adds your "thing" as a global member.
-          p
-            code-block-slotted(language="typescript")
-              template(v-slot:title) {{ example_code.bootstrap.filepath }}
-              template(v-slot:code) {{ example_code.bootstrap.contents }}
-        li
-          p Create your app file and have it use your global member.
-          p
-            code-block-slotted(language="typescript")
-              template(v-slot:title) {{ example_code.app.filepath }}
-              template(v-slot:code) {{ example_code.app.contents }}
-  div.row
-    div.col
-      hr
-      h2-hash Verification
-      ol
-        li Run your app.
-          code-block-slotted
-            template(v-slot:title) Terminal
-            template(v-slot:code)
-              | deno run app.ts
-          p You should see the following after you have run your app:
-          code-block-slotted
-            template(v-slot:title) Terminal
-            template(v-slot:code)
-              | Hello from MyThing!
-          p As you can see, <code>app.ts</code> knows nothing about <code>MyThing</code>. All it knows is that it should be registered as a global member and that it can use it without having to import it.
+  h2-hash Before You Get Started
+  p A global member is any "thing" you define the global member to be. It can be a service, an app, an array, a number, etc.
+  p You can add global members to the <code>Drash.Members</code> namespace via <code>Drash.addMember()</code>.
+  p Adding global members is helpful when you want to access a "thing" without having to remember where it's physically located. This prevents you from having to import like <code>import MyThing from "../path/../to/../some/../dir/my_thing.ts"</code>. Instead, you can add your "thing" as a global member (e.g., <code>Drash.addMember("MyThing", myThing);</code>) and access it like <code>Drash.Members.MyThing</code>.
+  p All global members are added to the <code>Drash.Members</code> namespace.
+  hr
+  folder-structure-end-state
+    | ▾ /path/to/your/project/
+    |   app.ts
+    |   bootstrap.ts
+    |   my_thing.ts
+  hr
+  h2-hash Steps
+  ol
+    li
+      p Create your "thing" to be added as a global member.
+      code-block(:title="example_code.my_thing.filepath" language="typescript")
+        | {{ example_code.my_thing.contents }}
+    li
+      p Create a file that adds your "thing" as a global member.
+      code-block(:title="example_code.bootstrap.filepath" language="typescript")
+        | {{ example_code.bootstrap.contents }}
+    li
+      p Create your app file and have it use your global member.
+      code-block(:title="example_code.app.filepath" language="typescript")
+        | {{ example_code.app.contents }}
+  hr
+  h2-hash Verification
+  ol
+    li
+      P Run your app.
+      code-block(title="Terminal")
+        | deno run app.ts
+      p You should see the following after you have run your app:
+      code-block(title="Terminal")
+        | Hello from MyThing!
+      p As you can see, <code>app.ts</code> knows nothing about <code>MyThing</code>. All it knows is that it should be registered as a global member and that it can use it without having to import it.
 </template>
